+++ drivers/androidx-sqlite-driver/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=androidx-sqlite-driver
+POM_NAME=SQLDelight Androidx SQLiteDriver
+POM_DESCRIPTION=Driver to support SQLDelight generated code using the Androidx SQLiteDriver (Android, iOS, Mac, Linux, and JVM Desktop)
