--- drivers/kmp-driver/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=kmp-driver
-POM_NAME=SQLDelight KMP Driver
-POM_DESCRIPTION=Driver to support SQLDelight generated code on KMP (Android, iOS, Mac, Linux, and JVM Desktop)
