+++ settings.gradle
@@ -41,8 +41,8 @@ include ':dialects:sqlite-3-33'
 include ':dialects:sqlite-3-35'
 include ':dialects:sqlite-3-38'
 include ':drivers:android-driver'
+include ':drivers:androidx-sqlite-driver'
 include ':drivers:jdbc-driver'
-include ':drivers:kmp-driver'
 include ':drivers:r2dbc-driver'
 include ':drivers:native-driver'
 include ':drivers:sqlite-driver'
